feat: create provider for recipes
