+++ src/index.js
@@ -2,12 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import './global.scss';
+import { RecipesProvider } from './context/Recipes';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
   <BrowserRouter>
-    <App />
+    <RecipesProvider>
+      <App />
+    </RecipesProvider>
   </BrowserRouter>, document.getElementById('root'),
 );
