+++ src/pages/MainRecipes/index.jsx
@@ -1,31 +1,14 @@
-import { useHistory } from 'react-router-dom';
-import React, { useEffect, useState } from 'react';
+import React, { useContext } from 'react';
+import { RecipesContext } from '../../context/Recipes';
 import Header from '../../components/Header';
-import getRecipes from '../../services/recipesData';
 import RecipesCardsGrid from '../../components/RecipesCardsGrid';
-import Categories from '../../components/Categories';
+import Categories from './components/Categories';
 import Carousel from '../../components/Carousel';
 
 import styles from './styles.module.scss';
 
 function MainRecipes() {
-  const { location: { pathname } } = useHistory();
-  const [recipes, setRecipes] = useState([]);
-  const [categories, setCategories] = useState([]);
-  const [titlePage, setTitlePage] = useState('');
-
-  useEffect(() => {
-    async function loadRecipes() {
-      const recipesLimit = 12;
-      const categoriesLimit = 5;
-      const results = await getRecipes(pathname);
-      setRecipes(results.list.slice(0, recipesLimit));
-      setCategories(results.categories.slice(0, categoriesLimit));
-      setTitlePage(results.titlePage);
-    }
-
-    loadRecipes();
-  }, [pathname]);
+  const { recipes, categories, titlePage } = useContext(RecipesContext);
 
   return (
     <div className={ styles.mainRecipes }>
